refactor: refactor amount calculations to use float64 type
- Change the type of the `amount` parameter in `CalculateProgress` from int64 to float64
- Update the calculation in `CalculateProgress` to directly use the new float64 amount type
- Modify the return type of `GetTotalSwapAmount` from int64 to float64
- Change the variable type for `totalAmount` in `GetTotalSwapAmount` to float64
- Update error handling in `getTokenDetails` to use `%w` for wrapping errors instead of `%v`

+++ entity/domain/core/biz/agg_task_test.go
@@ -125,7 +125,7 @@ func TestTask_CalculateProgress(t1 *testing.T) {
 		poolID    string
 	}
 	type args struct {
-		amount int64
+		amount float64
 	}
 	tests := []struct {
 		name   string
